Refactor Drawable.drawable into drawable
