--- docs/source/foundry.core.drawable.rst
@@ -1,21 +1,0 @@
-foundry.core.drawable package
-=============================
-
-Submodules
-----------
-
-foundry.core.drawable.Drawable module
--------------------------------------
-
-.. automodule:: foundry.core.drawable.Drawable
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: foundry.core.drawable
-   :members:
-   :undoc-members:
-   :show-inheritance:
