+++ foundry/core/drawable.py
